+++ src/main/java/svenhjol/charm/module/StrayImprovements.java
@@ -27,7 +27,7 @@ public class StrayImprovements extends MesonModule {
     public static boolean dropIce = true;
 
     @Config(name = "Maximum drops", description = "Maximum blue ice dropped when stray is killed.")
-    public static double maxDrops = 2;
+    public static int maxDrops = 2;
 
     @Module(description = "Strays spawn anywhere within their biome and have a chance to drop blue ice.", hasSubscriptions = true)
     public StrayImprovements() { }
